First Iteration of Lambda URL Checker
Why these changes are being introduced:
This is the first iteration of the Lambda Function URL Checker
CloudWatch Synthetic Canary. After experimenting with Python+Selenium,
I determined that there were too many issues with that model (mostly
due to the fact that the python-selenium runtime would not post metrics
to CloudWatch properly). So, this is a switch to Javascript + Node.js
which uses the javascript-node-puppeteer runtime in Synthetic Canaries.

How this addresses that need:
* Create an index.js file in the nodejs directory in the project folder
in the repository with all the code to check a Lambda Function URL's
HTTP Response Code and write that data point to CloudWatch metrics
* Create a package.json with the necessary dev dependencies for local
testing with `jest`
* Create a zipmanifest.txt file to track that list of files that will
need to be zipped up and stored in S3 so that the Synthetics Canary can
build the function in AWS

Side effects of this change:
None. There is no deployment methodology in this commit.