Test Suite for Lambda Function URL Checker
Why these changes are being introduced:
This is the suite of tests for the Javascript + Node.js application that
uses `jest`. This also includes a Makefile with dependency instal
commands, test commands, and a dev-build command to push a .zip of the
code to S3.

How this addresses that need:
* Create mocks for the Synthetics and SyntheticsLogger so that we can
run tests
* Create tests for the majority of the code in the handler, main, and
_checkUrl functions
* Create a Makefile for re-usable commands for dependency installation,
tests, and a dev deploy to AWS
* Create a CI GHA workflow that will run the tests in GitHub

Side effects of this change:
None. There is no automated deploy in GitHub Actions, just testing.