Reorganize For Upcoming GHA Workflows
Why these changes are being introduced:
Some testing was done in a separate feature branch to work out the
details of how we will handle automated deployment using GitHub Actions.
It turns out that a simpler repository folder structure will make this
a little easier.

How this addresses that need:
* Remove the `nodejs` subfolder from the lambda_function_url_checker
and flatten the files/folders slightly
* Update the test code to find all the dependencies correctly in the
new folder structure
* Update the package.json to reflect the new file/folder structure
* Create a .runtime file for the developer to designate the expected
runtime for the application in CloudWatch Synthetic Canary
* Create a Makefile in the application folder for application-specific
commands/targets
* Update the main Makefile to simply call subfolder Makefiles instead
* Clean up the CI workflow file slightly (addressing some typos) and
set the trigger for PRs to main (and manual triggers in the UI)

Side effects of this change:
None.

Related Jira Tickets:
* https://mitlibraries.atlassian.net/browse/IN-1532