Add Dev Deployment Workflow
Why these changes are being introduced:
This is the first pass and a deployment workflow in GitHub Actions to
push the ZIP file to the S3 bucket and set SSM Parameter Store
parameters so that the Terraform repository using this function can
set the correct values.

How this is being implemented:
* Create a new dev-deploy.yml workflow
* Create a new generate-matrix.sh bash script to generate the matrix of
deployment jobx

More details on the dev-deploy workflow:
* Use the generate-matrix.sh script to generate the list of top level
folders that have changes
* For each folder list in the matrix, run a deployment job to zip the
application, push the ZIP to the shared-files S3 bucket, and update two
SSM Parameter Store parameters with the key to the ZIP and the expected
runtime in CloudWatch Synthetic Canaries

More details on the generate-matrix.sh script:
* This was written with assistance from GitHub Copilot (but is mostly
an update to the matrix method that was first used by DLS in the
archived mitlib-terraform repository)
* It was moved to a shell script to allow for easier local testing to
ensure that it was actually doing what was expected without having to
run it multiple times in GitHub as an Actions
* It creates a list of top-level folders that have changes (from a
git diff), accounting for the various ways the Action could be running
(first push on new branch, triggered by a PR, triggered manually in the
GitHub UI)
* It exports the list to the GITHUB_OUTPUT environment so that the
Action can use this information to build the matrix to run for as many
folders have updated code

Side effects:
None.

Related Jira tickets:
* https://mitlibraries.atlassian.net/browse/IN-1531