Create Stage & Prod GHA Stubs
Why these changes are being introduced:
To make testing easier in the near future for the stage & prod
deployment workflows, we create stub workflows for now so that they
will exist in the `main` branch and allow for easier testing when we
finalize the steps.

How this addresses that need:
* Create a stage-deploy.yml workflow with just a stub set of steps
* Create a prod-deploy.yml workflow with just a stub set of steps

These two workflows will generate a matrix and also attempt to log in
to AWS with OIDC but will not create any ZIP files.

Side effects of this change:
None.