Update Workflows
Why these changes are being made:
The first attempt at the automated deployment workflows failed when it
got to the point of deploying in stage & prod environments (the script
failed to create the matrix list). The changes here do two things.
First, there are updates to the GHA workflows to handle the delivery
of the ZIP file to the S3 bucket in the correct AWS Account. Second,
there are updates to the shell script that generates the matrix JSON
object to determine which application should be updated.

How these changes are implemented:
* Update all three GHA workflows to the latest actions/checkout and
aws-actions/configure-aws-credentials
* Update all three GHA workflows to pull the last 10 commits when
running actions/checkout (to ensure that we have enough commit history
to handle the git diff)
* Remove the workflow_dispatch trigger (since that trigger does not
play nice with any sort of git diff approach)
* Add a conditional so that the workflows will not run if triggered by
dependabot
* Add environment variables that are pulled from the github context
so that the shell script has the right git history information to
generate the correct git diff
* Update the prod-deploy GHA workflow to match the deployment process
used in other repositories: copy the ZIP from stage to prod
* Update the script with additional documentation
* Update the script to account for the significantly different GHA
context for a published release tag workflow
* Update the script to allow for better local testing

WIP: exclude dependabot